[bug-fix][#22]props bug fix: props로 icon념겨주고 children받음
